freelance_nlqa/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
openai>=0.27.0
python-dotenv>=0.21.0
rich>=12.0.0
tabulate>=0.8.9
requests>=2.28
pytest>=7.0
langchain>=0.0.177
sqlite-utils>=3.36
tqdm>=4.64.0
